Game table fixes (#72)
* Fix crashes in GameTableManager under certian conditions.
Better logging to indicate which file failed to load.
Load any language files present, not just english. In the future, when we ship client language to the server, we'll dynamically swap the table in use for lookups. For now, it's forced to english like it was.

* Unwrap the inner continuations too.

* Moved text searches out of GameTableManager and into it's own class.
Added multi-language support (Not currently wired up, always selects english)