Message handler for client packet 0x23C
This packet a couple unrelated sets of data.
First is a random number the client generates and its seed value, second is a set of statistics on the watchdog task which gives an indication of how much load the client's computer is under.
There is currently nothing done with the data when received but the handler is there to stop the unknown packet log messages.