Addition of server command to allow changing client's entity send update interval
Figured out what the unknown parameter is server message 0xF1 (ServerInstanceSettings) does.
It sets the client's maximum send update interval for the player's unit.
This same parameter can be changed at any time by sending server message 0x70 (ServerClientEntityUpdateInterval).

Based on captures, the server would always set this parameter as 125 or 126ms and would periodically change it but it is not clear why.