Renamed field to ClientSendUpdateInterval
It was unclear what "max" send update interval means. It suggests that the client will always send updates every X interval however the entity updates from the client are also dependent on the user moving the character. The client does not send entity updates when a unit is stationary.
This parameter sets the interval at which updates are sent when the activity threshold determined by the client is met. It seems wise to get rid of the word max as it gives notion that the send rate varies and must be limited which is not the case.