Addition of server command to allow changing client's entity send update interval (#474)
* Addition of server command to allow changing client's entity send update interval

Figured out what the unknown parameter is server message 0xF1 (ServerInstanceSettings) does.
It sets the client's maximum send update interval for the player's unit.
This same parameter can be changed at any time by sending server message 0x70 (ServerClientEntityUpdateInterval).

Based on captures, the server would always set this parameter as 125 or 126ms and would periodically change it but it is not clear why.

* Renamed field to ClientSendUpdateInterval

It was unclear what "max" send update interval means. It suggests that the client will always send updates every X interval however the entity updates from the client are also dependent on the user moving the character. The client does not send entity updates when a unit is stationary.
This parameter sets the interval at which updates are sent when the activity threshold determined by the client is met. It seems wise to get rid of the word max as it gives notion that the send rate varies and must be limited which is not the case.

* Added comment as to when this message is valid to send